moved apps/server -> server
